+++ tlv/src/profiles/der.c
@@ -6,7 +6,8 @@
 #include "der_values_internal.h"
 #include <string.h>
 
-const tlv_der_limits_t tlv_der_default_limits = {32, 16 * 1024 * 1024, 16 * 1024 * 1024, 100000};
+const tlv_der_limits_t tlv_der_default_limits = {32, (size_t)16 * 1024 * 1024,
+                                                 (size_t)16 * 1024 * 1024, 100000};
 
 static tlv_result_t fail(tlv_result_t rc, size_t offset, size_t* error_offset) {
     if (error_offset) *error_offset = offset;
